pyproject.toml
[project]
name = "power-peak-shift"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit",
    "pandas",
    "numpy",
    "matplotlib",
    "japanize-matplotlib",
    "openpyxl",
    "google-cloud-storage",
]
